refactor the base class
